Update the CI to use fastlane and deliver aab
This new way to deliver will speed up the download on the Google Play Store and is the up do date way to release new apps. To achieve that, I use the fastlane tool to automatize the process, based on a custom docker image containing the ruby dependencies. In addition, I upgraded dependency versions and sdk versions, and clean useless code